requirements.txt
streamlit
google-generative-ai
pillow
pandas